+++ crates/ruma-events/src/room/avatar.rs
@@ -3,6 +3,8 @@
 //! [`m.room.avatar`]: https://spec.matrix.org/latest/client-server-api/#mroomavatar
 
 use js_int::UInt;
+#[cfg(feature = "unstable-msc2448")]
+use ruma_common::serde::Base64;
 use ruma_common::OwnedMxcUri;
 use ruma_macros::EventContent;
 use serde::{Deserialize, Serialize};
@@ -69,6 +71,14 @@ pub struct ImageInfo {
     #[cfg(feature = "unstable-msc2448")]
     #[serde(rename = "xyz.amorgan.blurhash", skip_serializing_if = "Option::is_none")]
     pub blurhash: Option<String>,
+
+    /// The [ThumbHash](https://evanw.github.io/thumbhash/) for this image.
+    ///
+    /// This uses the unstable prefix in
+    /// [MSC2448](https://github.com/matrix-org/matrix-spec-proposals/pull/2448).
+    #[cfg(feature = "unstable-msc2448")]
+    #[serde(rename = "xyz.amorgan.thumbhash", skip_serializing_if = "Option::is_none")]
+    pub thumbhash: Option<Base64>,
 }
 
 impl ImageInfo {